migration to native orientdb api;
