+++ orientdb/pom.xml
@@ -66,6 +66,11 @@
                 <groupId>com.orientechnologies</groupId>
                 <artifactId>orientdb-core</artifactId>
             </dependency>
+            <dependency>
+                <groupId>com.orientechnologies</groupId>
+                <artifactId>orientdb-client</artifactId>
+                <version>2.2.12</version>
+            </dependency>
         <dependency>
             <groupId>org.parboiled</groupId>
             <artifactId>parboiled-java</artifactId>
@@ -157,6 +162,6 @@
             <groupId>org.jboss.byteman</groupId>
             <artifactId>byteman-install</artifactId>
             <scope>test</scope>
-        </dependency>
+        </dependency>        
     </dependencies>
 </project>
