+++ orientdb/src/main/java/org/hibernate/ogm/datastore/orientdb/OrientDBDialect.java
@@ -7,8 +7,6 @@
 package org.hibernate.ogm.datastore.orientdb;
 
 import java.io.Serializable;
-import java.sql.Connection;
-import java.sql.PreparedStatement;
 import java.sql.SQLException;
 import java.util.Arrays;
 import java.util.Collection;
@@ -22,13 +20,12 @@ import java.util.Map;
 import org.hibernate.StaleObjectStateException;
 import org.hibernate.engine.spi.SessionFactoryImplementor;
 import org.hibernate.ogm.datastore.orientdb.constant.OrientDBConstant;
-import org.hibernate.ogm.datastore.orientdb.constant.OrientDBMapping;
 import org.hibernate.ogm.datastore.orientdb.dialect.impl.OrientDBAssociationQueries;
 import org.hibernate.ogm.datastore.orientdb.dialect.impl.OrientDBAssociationSnapshot;
 import org.hibernate.ogm.datastore.orientdb.dialect.impl.OrientDBEntityQueries;
 import org.hibernate.ogm.datastore.orientdb.dialect.impl.OrientDBTupleAssociationSnapshot;
 import org.hibernate.ogm.datastore.orientdb.dialect.impl.OrientDBTupleSnapshot;
-import org.hibernate.ogm.datastore.orientdb.dialect.impl.ResultSetTupleIterator;
+import org.hibernate.ogm.datastore.orientdb.dialect.impl.ODocumentListTupleIterator;
 import org.hibernate.ogm.datastore.orientdb.dto.GenerationResult;
 import org.hibernate.ogm.datastore.orientdb.impl.OrientDBDatastoreProvider;
 import org.hibernate.ogm.datastore.orientdb.impl.OrientDBSchemaDefiner;
@@ -81,10 +78,12 @@ import org.hibernate.persister.collection.CollectionPersister;
 import org.hibernate.persister.entity.EntityPersister;
 import org.hibernate.type.Type;
 
+import com.orientechnologies.orient.core.db.document.ODatabaseDocumentTx;
 import com.orientechnologies.orient.core.db.record.ridbag.ORidBag;
 import com.orientechnologies.orient.core.exception.OConcurrentModificationException;
 import com.orientechnologies.orient.core.id.ORecordId;
 import com.orientechnologies.orient.core.record.impl.ODocument;
+import org.hibernate.ogm.datastore.orientdb.utils.QueryUtil;
 
 /**
  * Implementation of dialect for OrientDB
@@ -103,7 +102,7 @@ import com.orientechnologies.orient.core.record.impl.ODocument;
  * @author Sergey Chernolyas &lt;sergey.chernolyas@gmail.com&gt;
  */
 public class OrientDBDialect extends BaseGridDialect
-		implements QueryableGridDialect<String>, SessionFactoryLifecycleAwareDialect, IdentityColumnAwareGridDialect {
+implements QueryableGridDialect<String>, SessionFactoryLifecycleAwareDialect, IdentityColumnAwareGridDialect {
 
 	private static final long serialVersionUID = 1L;
 	private static final Log log = LoggerFactory.getLogger();
@@ -164,9 +163,9 @@ public class OrientDBDialect extends BaseGridDialect
 		Tuple tuple = tuplePointer.getTuple();
 		log.debugf( "insertOrUpdateTuple:EntityKey: %s ; tupleContext: %s ; tuple: %s ; thread: %s",
 				key, tupleContext, tuple, Thread.currentThread().getName() );
-		Connection connection = provider.getConnection();
+		ODatabaseDocumentTx db = provider.getConnection();
 		OrientDBTupleSnapshot snapshot = (OrientDBTupleSnapshot) tuple.getSnapshot();
-		boolean existsInDB = EntityKeyUtil.existsPrimaryKeyInDB( connection, key );
+		boolean existsInDB = EntityKeyUtil.existsPrimaryKeyInDB( db, key );
 		QueryType queryType = QueryTypeDefiner.define( existsInDB, snapshot.isNew() );
 		log.debugf( "insertOrUpdateTuple: snapshot.isNew(): %b ,snapshot.isEmpty(): %b; exists in DB: %b; query type: %s ",
 				snapshot.isNew(), snapshot.isEmpty(), existsInDB, queryType );
@@ -192,17 +191,18 @@ public class OrientDBDialect extends BaseGridDialect
 
 		try {
 			log.debugf( "insertOrUpdateTuple:Key: %s; Query: %s; ", key, queryBuffer );
-			PreparedStatement pstmt = connection.prepareStatement( queryBuffer.toString() );
-			int updateCount = pstmt.executeUpdate();
-			log.debugf( "insertOrUpdateTuple:Key: %s ;inserted or updated: %d ", key, updateCount );
-			if ( updateCount == 0 && queryType.equals( QueryType.UPDATE ) ) {
-				// primary key was in DB .... but during prepare query someone remove it from DB.
-				Integer currentVersion = (Integer) snapshot.get( OrientDBConstant.SYSTEM_VERSION );
-				throw log.versionNotActual( key, currentVersion );
-			}
-		}
-		catch (SQLException sqle) {
-			throw log.cannotExecuteQuery( queryBuffer.toString(), sqle );
+			QueryUtil.executeNativeQuery( db, queryBuffer );
+			// PreparedStatement pstmt = connection.prepareStatement( queryBuffer.toString() );
+			// int updateCount = pstmt.executeUpdate();
+			// log.debugf( "insertOrUpdateTuple:Key: %s ;inserted or updated: %d ", key, updateCount );
+			// if ( updateCount == 0 && queryType.equals( QueryType.UPDATE ) ) {
+			// primary key was in DB .... but during prepare query someone remove it from DB.
+			// Integer currentVersion = (Integer) snapshot.get( OrientDBConstant.SYSTEM_VERSION );
+			// throw log.versionNotActual( key, currentVersion );
+			// }
+			/*
+			 * } catch (SQLException sqle) { throw log.cannotExecuteQuery( queryBuffer.toString(), sqle );
+			 */
 		}
 		catch (OConcurrentModificationException cme) {
 			throw new StaleObjectStateException( key.getTable(), (Serializable) EntityKeyUtil.generatePrimaryKeyPredicate( key ) );
@@ -216,7 +216,7 @@ public class OrientDBDialect extends BaseGridDialect
 
 		String dbKeyName = entityKeyMetadata.getColumnNames()[0];
 		Long dbKeyValue = null;
-		Connection connection = provider.getConnection();
+
 		String query = null;
 
 		if ( dbKeyName.equals( OrientDBConstant.SYSTEM_RID ) ) {
@@ -226,7 +226,7 @@ public class OrientDBDialect extends BaseGridDialect
 		else {
 			// use business key. get new id from sequence
 			String seqName = OrientDBSchemaDefiner.generateSeqName( entityKeyMetadata.getTable(), dbKeyName );
-			dbKeyValue = (Long) SequenceUtil.getNextSequenceValue( connection, seqName );
+			dbKeyValue = (Long) SequenceUtil.getNextSequenceValue( provider.getConnection(), seqName );
 			tuple.put( dbKeyName, dbKeyValue );
 		}
 		GenerationResult result = INSERT_QUERY_GENERATOR.generate( entityKeyMetadata.getTable(), tuple, true,
@@ -234,38 +234,38 @@ public class OrientDBDialect extends BaseGridDialect
 		query = result.getExecutionQuery();
 
 		log.debugf( "insertTuple: insertQuery: %s ", result.getExecutionQuery() );
-		try {
-			PreparedStatement pstmt = connection.prepareStatement( result.getExecutionQuery() );
-			log.debugf( "insertTuple:Key: %s (%s) ;inserted or updated: %d ", dbKeyName, dbKeyValue, pstmt.executeUpdate() );
-		}
-		catch (SQLException sqle) {
-			throw log.cannotExecuteQuery( query, sqle );
-		}
+		QueryUtil.executeNativeQuery( provider.getConnection(), result.getExecutionQuery() );
+		/*
+		 * try { PreparedStatement pstmt = connection.prepareStatement( result.getExecutionQuery() ); log.debugf(
+		 * "insertTuple:Key: %s (%s) ;inserted or updated: %d ", dbKeyName, dbKeyValue, pstmt.executeUpdate() ); } catch
+		 * (SQLException sqle) { throw log.cannotExecuteQuery( query, sqle ); }
+		 */
 	}
 
 	@Override
 	public void removeTuple(EntityKey key, TupleContext tupleContext) {
 		log.debugf( "removeTuple:EntityKey: %s ; tupleContext %s ; current thread: %s",
 				key, tupleContext, Thread.currentThread().getName() );
-		Connection connection = provider.getConnection();
+		ODatabaseDocumentTx db = provider.getConnection();
 		StringBuilder queryBuffer = new StringBuilder( 100 );
-		try {
-			queryBuffer.append( "DELETE VERTEX " ).append( key.getTable() ).append( " where " ).append( EntityKeyUtil.generatePrimaryKeyPredicate( key ) );
-			log.debugf( "removeTuple:Key: %s. query: %s ", key, queryBuffer );
-			PreparedStatement pstmt = connection.prepareStatement( queryBuffer.toString() );
-			log.debugf( "removeTuple:Key: %s. remove: %s", key, pstmt.executeUpdate() );
-		}
-		catch (SQLException e) {
-			throw log.cannotExecuteQuery( queryBuffer.toString(), e );
-		}
+		// try {
+		queryBuffer.append( "DELETE VERTEX " ).append( key.getTable() ).append( " where " ).append( EntityKeyUtil.generatePrimaryKeyPredicate( key ) );
+		log.debugf( "removeTuple:Key: %s. query: %s ", key, queryBuffer );
+		QueryUtil.executeNativeQuery( db, queryBuffer );
+		// PreparedStatement pstmt = db.prepareStatement( queryBuffer.toString() );
+		// log.debugf( "removeTuple:Key: %s. remove: %s", key, pstmt.executeUpdate() );
+		// }
+		// catch (SQLException e) {
+		// throw log.cannotExecuteQuery( queryBuffer.toString(), e );
+		// }
 	}
 
 	@Override
 	public Association getAssociation(AssociationKey associationKey, AssociationContext associationContext) {
 		log.debugf( "getAssociation:AssociationKey: %s ; AssociationContext: %s", associationKey, associationContext );
 		EntityKey entityKey = associationKey.getEntityKey();
-		Connection connection = provider.getConnection();
-		boolean existsPrimaryKey = EntityKeyUtil.existsPrimaryKeyInDB( connection, entityKey );
+		ODatabaseDocumentTx db = provider.getConnection();
+		boolean existsPrimaryKey = EntityKeyUtil.existsPrimaryKeyInDB( db, entityKey );
 		if ( !existsPrimaryKey ) {
 			// Entity now extists
 			return ASSOCIATION_NOT_FOUND;
@@ -277,9 +277,9 @@ public class OrientDBDialect extends BaseGridDialect
 
 	private Map<RowKey, Tuple> createAssociationMap(AssociationKey associationKey, AssociationContext associationContext) {
 		log.debugf( "createAssociationMap:AssociationKey: %s ; AssociationContext: %s", associationKey, associationContext );
-		Connection connection = provider.getConnection();
+		ODatabaseDocumentTx db = provider.getConnection();
 		List<Map<String, Object>> relationships = entityQueries.get( associationKey.getEntityKey().getMetadata() )
-				.findAssociation( connection, associationKey, associationContext );
+				.findAssociation( db, associationKey, associationContext );
 
 		Map<RowKey, Tuple> tuples = new LinkedHashMap<>();
 
@@ -365,9 +365,9 @@ public class OrientDBDialect extends BaseGridDialect
 	private void putAssociationOperation(Association association, AssociationKey associationKey, AssociationOperation action,
 			AssociatedEntityKeyMetadata associatedEntityKeyMetadata) {
 		log.debugf( "putAssociationOperation: : action: %s ; metadata: %s; association:%s", action, associationKey.getMetadata(), association );
-		Connection connection = provider.getConnection();
+		ODatabaseDocumentTx db = provider.getConnection();
 		if ( associationQueries.containsKey( associationKey.getMetadata() ) ) {
-			List<Map<String, Object>> relationship = associationQueries.get( associationKey.getMetadata() ).findRelationship( connection,
+			List<Map<String, Object>> relationship = associationQueries.get( associationKey.getMetadata() ).findRelationship( db,
 					associationKey, action.getKey() );
 			if ( relationship.isEmpty() ) {
 				// create
@@ -391,13 +391,7 @@ public class OrientDBDialect extends BaseGridDialect
 				associationKey.getMetadata(), associationRow, associatedEntityKeyMetadata );
 		GenerationResult result = INSERT_QUERY_GENERATOR.generate( associationKey.getTable(), associationRow, false, Collections.<String>emptySet() );
 		log.debugf( "createRelationshipWithNode: query: %s", result.getExecutionQuery() );
-		try {
-			PreparedStatement pstmt = provider.getConnection().prepareStatement( result.getExecutionQuery() );
-			log.debugf( "createRelationshipWithNode: execute insert query: %d", pstmt.executeUpdate() );
-		}
-		catch (SQLException sqle) {
-			throw log.cannotExecuteQuery( result.getExecutionQuery(), sqle );
-		}
+		QueryUtil.executeNativeQuery( provider.getConnection(), result.getExecutionQuery() );
 	}
 
 	private void updateRelationshipWithNode(AssociationKey associationKey, Tuple associationRow,
@@ -407,13 +401,8 @@ public class OrientDBDialect extends BaseGridDialect
 		if ( AssociationKind.EMBEDDED_COLLECTION.equals( associationKey.getMetadata().getAssociationKind() ) ) {
 			GenerationResult result = UPDATE_QUERY_GENERATOR.generate( associationKey, associationRow );
 			log.debugf( "updateRelationshipWithNode: query: %s", result.getExecutionQuery() );
-			try {
-				PreparedStatement pstmt = provider.getConnection().prepareStatement( result.getExecutionQuery() );
-				log.debugf( "updateRelationshipWithNode: execute update query: %d", pstmt.executeUpdate() );
-			}
-			catch (SQLException sqle) {
-				throw log.cannotExecuteQuery( result.getExecutionQuery(), sqle );
-			}
+			QueryUtil.executeNativeQuery( provider.getConnection(), result.getExecutionQuery() );
+			// log.debugf( "updateRelationshipWithNode: execute update query: %d", pstmt.executeUpdate() );
 		}
 		else {
 			log.debugf( "updateRelationshipWithNode: update  association  not needs!" );
@@ -429,12 +418,12 @@ public class OrientDBDialect extends BaseGridDialect
 	public Number nextValue(NextValueRequest request) {
 		log.debugf( "NextValueRequest: %s", request );
 		long nextValue = 0;
-		Connection connection = provider.getConnection();
+		ODatabaseDocumentTx db = provider.getConnection();
 		IdSourceType type = request.getKey().getMetadata().getType();
 		switch ( type ) {
 			case SEQUENCE:
 				String seqName = request.getKey().getMetadata().getName();
-				nextValue = SequenceUtil.getNextSequenceValue( connection, seqName );
+				nextValue = SequenceUtil.getNextSequenceValue( db, seqName );
 				break;
 			case TABLE:
 				String seqTableName = "sequences";
@@ -443,7 +432,7 @@ public class OrientDBDialect extends BaseGridDialect
 				String pkColumnValue = (String) request.getKey().getColumnValue();
 				log.debugf( "seqTableName:%s, pkColumnName:%s, pkColumnValue:%s, valueColumnName:%s",
 						seqTableName, pkColumnName, pkColumnValue, valueColumnName );
-				nextValue = SequenceUtil.getNextTableValue( connection, seqTableName, pkColumnName, pkColumnValue, valueColumnName,
+				nextValue = SequenceUtil.getNextTableValue( db, seqTableName, pkColumnName, pkColumnValue, valueColumnName,
 						request.getInitialValue(), request.getIncrement() );
 		}
 		log.debugf( "nextValue: %d", nextValue );
@@ -465,57 +454,38 @@ public class OrientDBDialect extends BaseGridDialect
 	 */
 
 	@SuppressWarnings("unchecked")
-	private PreparedStatement prepareStatement(BackendQuery<String> backendQuery, QueryParameters queryParameters) throws SQLException {
+	private List<ODocument> executeNativeQueryWithParams(BackendQuery<String> backendQuery, QueryParameters queryParameters) {
 		Map<String, Object> parameters = getNamedParameterValuesConvertedByGridType( queryParameters );
 		log.debugf( "prepareStatement: parameters: %s ; ",
 				parameters.keySet() );
-		String nativeQuery = backendQuery.getQuery();
+		String nativeQueryTemplate = backendQuery.getQuery();
 		log.debugf( "prepareStatement: nativeQuery: %s ; metadata: %s",
 				backendQuery.getQuery(), backendQuery.getSingleEntityMetadataInformationOrNull() );
-		PreparedStatement pstmt = provider.getConnection().prepareStatement( nativeQuery );
-		int paramIndex = 1;
+		Map<String, Object> queryParams = new HashMap<>();
 		for ( Map.Entry<String, TypedGridValue> entry : queryParameters.getNamedParameters().entrySet() ) {
 			String key = entry.getKey();
 			TypedGridValue value = entry.getValue();
 			log.debugf( "prepareStatement: key: %s ; type: %s ; value: %s; type class: %s ",
 					key, value.getType(), value.getValue(), value.getType().getReturnedClass() );
-			try {
-				if ( OrientDBMapping.SIMPLE_VALUE_SETTER_MAP.containsKey( value.getType() ) ) {
-					OrientDBMapping.SIMPLE_VALUE_SETTER_MAP.get( value.getType() ).setValue( pstmt, paramIndex, value.getValue() );
-				}
-				else {
-					throw new UnsupportedOperationException( "Type " + value.getType() + " is not supported!" );
-				}
-
-			}
-			catch (SQLException sqle) {
-				throw log.cannotSetValueForParameter( paramIndex, sqle );
-			}
-			paramIndex++;
+			queryParams.put( key, value.getValue() );
 		}
-		return pstmt;
+		return QueryUtil.executeNativeQueryWithParams( provider.getConnection(), nativeQueryTemplate, queryParams );
 	}
 
 	@Override
 	public int executeBackendUpdateQuery(BackendQuery<String> query, QueryParameters queryParameters, TupleContext tupleContext) {
-		try {
-			PreparedStatement pstmt = prepareStatement( query, queryParameters );
-			return pstmt.executeUpdate();
-		}
-		catch (SQLException e) {
-			throw log.cannotExecuteQuery( query.getQuery(), e );
-		}
+		List<ODocument> documents = executeNativeQueryWithParams( query, queryParameters );
+		// @todo fix this
+		return -1;
+
 	}
 
 	@Override
 	public ClosableIterator<Tuple> executeBackendQuery(BackendQuery<String> query, QueryParameters queryParameters, TupleContext tupleContext) {
-		try {
-			PreparedStatement pstmt = prepareStatement( query, queryParameters );
-			return new ResultSetTupleIterator( pstmt.executeQuery() );
-		}
-		catch (SQLException e) {
-			throw log.cannotExecuteQuery( query.getQuery(), e );
-		}
+
+		List<ODocument> documents = executeNativeQueryWithParams( query, queryParameters );
+		return new ODocumentListTupleIterator( documents );
+
 	}
 
 	/**
