+++ orientdb/src/main/java/org/hibernate/ogm/datastore/orientdb/dialect/impl/OrientDBEntityQueries.java
@@ -7,22 +7,15 @@
 package org.hibernate.ogm.datastore.orientdb.dialect.impl;
 
 import java.math.BigDecimal;
-import java.sql.Connection;
-import java.sql.ResultSet;
-import java.sql.ResultSetMetaData;
-import java.sql.SQLException;
-import java.sql.Statement;
 import java.util.LinkedHashMap;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
-import java.util.Map.Entry;
 
 import org.hibernate.ogm.datastore.orientdb.constant.OrientDBConstant;
 import org.hibernate.ogm.datastore.orientdb.logging.impl.Log;
 import org.hibernate.ogm.datastore.orientdb.logging.impl.LoggerFactory;
 import org.hibernate.ogm.datastore.orientdb.utils.EntityKeyUtil;
-import org.hibernate.ogm.datastore.orientdb.utils.ODatabaseDocumentUtil;
 import org.hibernate.ogm.dialect.spi.AssociationContext;
 import org.hibernate.ogm.model.key.spi.AssociationKey;
 import org.hibernate.ogm.model.key.spi.EntityKey;
@@ -67,73 +60,56 @@ public class OrientDBEntityQueries extends QueriesBase {
 	 * @return the corresponding node
 	 */
 
-	public Map<String, Object> findEntity(Connection connection, EntityKey entityKey) {
+	public Map<String, Object> findEntity(ODatabaseDocumentTx db, EntityKey entityKey) {
 		Map<String, Object> dbValues = new LinkedHashMap<>();
 		StringBuilder query = new StringBuilder( "select from " );
-		//try {
-			//Statement stmt = connection.createStatement();
-			if ( entityKey.getColumnNames().length == 1 && entityKey.getColumnValues()[0] instanceof ORecordId ) {
-				// search by @rid
-				ORecordId rid = (ORecordId) entityKey.getColumnValues()[0];
-				query.append( rid );
-			}
-			else {
-				// search by business key
-				log.debugf( "column names: %s", Arrays.asList( entityKey.getColumnNames() ) );
-				query.append( entityKey.getTable() ).append( " WHERE " ).append( EntityKeyUtil.generatePrimaryKeyPredicate( entityKey ) );
+		// try {
+		// Statement stmt = connection.createStatement();
+		if ( entityKey.getColumnNames().length == 1 && entityKey.getColumnValues()[0] instanceof ORecordId ) {
+			// search by @rid
+			ORecordId rid = (ORecordId) entityKey.getColumnValues()[0];
+			query.append( rid );
+		}
+		else {
+			// search by business key
+			log.debugf( "column names: %s", Arrays.asList( entityKey.getColumnNames() ) );
+			query.append( entityKey.getTable() ).append( " WHERE " ).append( EntityKeyUtil.generatePrimaryKeyPredicate( entityKey ) );
 
+		}
+		log.debugf( "find entiry query: %s", query.toString() );
+		List<ODocument> documents = QueryUtil.executeNativeQuery( db, query );
+		if ( documents.isEmpty() ) {
+			log.debugf( " entity by primary key %s not found!", entityKey );
+			return null;
+		}
+		ODocument currentDocument = documents.get( 0 );
+		dbValues = currentDocument.toMap();
+		for ( String fieldName : dbValues.keySet() ) {
+			Object fieldValue = dbValues.get( fieldName );
+			if ( fieldValue instanceof ODocument ) {
+				dbValues.remove( fieldName );
+				dbValues.putAll( ODocumentUtil.extractNamesTree( fieldName, (ODocument) fieldValue ) );
 			}
-			log.debugf( "find entiry query: %s", query.toString() );
-			List<ODocument> documents = QueryUtil.executeNativeQuery( ODatabaseDocumentUtil.getDatabase( connection ), query );
-			if (documents.isEmpty()) {
-				log.debugf( " entity by primary key %s not found!", entityKey );
-				return null;
-			}
-			ODocument currentDocument  = documents.get( 0 );
-			dbValues = currentDocument.toMap();
-			for (String fieldName : dbValues.keySet()) {
-				Object fieldValue = dbValues.get( fieldName );
-				if ( fieldValue instanceof ODocument ) {
-					dbValues.remove( fieldName );
-					dbValues.putAll( ODocumentUtil.extractNamesTree( fieldName, (ODocument) fieldValue ) );
-				}
-				
-			}
-			reCastValues( dbValues );
-			/*
-			
-			ResultSet rs = stmt.executeQuery( query.toString() );
-			if ( rs.next() ) {
-				ResultSetMetaData metadata = rs.getMetaData();
-				for ( String systemField : OrientDBConstant.SYSTEM_FIELDS ) {
-					dbValues.put( systemField, rs.getObject( systemField ) );
-				}
-				for ( int i = 0; i < rs.getMetaData().getColumnCount(); i++ ) {
-					int dbFieldNo = i + 1;
-					String dbColumnName = metadata.getColumnName( dbFieldNo );
-					Object dbValue = rs.getObject( dbColumnName );
-					log.debugf( "%d dbColumnName: %s; dbValue class: %s; dbvalue: %s", i, dbColumnName, ( dbValue != null ? dbValue.getClass() : null ),
-							dbValue );
-					log.debugf( "%d dbColumnName: %s ; sql type: %s", i, dbColumnName, rs.getMetaData().getColumnTypeName( dbFieldNo ) );
-					dbValues.put( dbColumnName, dbValue );
-					if ( dbValue instanceof ODocument ) {
-						dbValues.remove( dbColumnName );
-						dbValues.putAll( ODocumentUtil.extractNamesTree( dbColumnName, (ODocument) dbValue ) );
-					}
-				}
-				reCastValues( dbValues );
-				log.debugf( " entity values from db:  %s", dbValues );
-			}
-			else {
-				log.debugf( " entity by primary key %s not found!", entityKey );
-				return null;
-			} */
-			return dbValues;
-		/*}
-		catch (SQLException sqle) {
-			throw log.cannotExecuteQuery( query.toString(), sqle );
 
-		} */
+		}
+		reCastValues( dbValues );
+		/*
+		 * ResultSet rs = stmt.executeQuery( query.toString() ); if ( rs.next() ) { ResultSetMetaData metadata =
+		 * rs.getMetaData(); for ( String systemField : OrientDBConstant.SYSTEM_FIELDS ) { dbValues.put( systemField,
+		 * rs.getObject( systemField ) ); } for ( int i = 0; i < rs.getMetaData().getColumnCount(); i++ ) { int
+		 * dbFieldNo = i + 1; String dbColumnName = metadata.getColumnName( dbFieldNo ); Object dbValue = rs.getObject(
+		 * dbColumnName ); log.debugf( "%d dbColumnName: %s; dbValue class: %s; dbvalue: %s", i, dbColumnName, ( dbValue
+		 * != null ? dbValue.getClass() : null ), dbValue ); log.debugf( "%d dbColumnName: %s ; sql type: %s", i,
+		 * dbColumnName, rs.getMetaData().getColumnTypeName( dbFieldNo ) ); dbValues.put( dbColumnName, dbValue ); if (
+		 * dbValue instanceof ODocument ) { dbValues.remove( dbColumnName ); dbValues.putAll(
+		 * ODocumentUtil.extractNamesTree( dbColumnName, (ODocument) dbValue ) ); } } reCastValues( dbValues );
+		 * log.debugf( " entity values from db:  %s", dbValues ); } else { log.debugf(
+		 * " entity by primary key %s not found!", entityKey ); return null; }
+		 */
+		return dbValues;
+		/*
+		 * } catch (SQLException sqle) { throw log.cannotExecuteQuery( query.toString(), sqle ); }
+		 */
 	}
 
 	private void reCastValues(Map<String, Object> map) {
@@ -166,7 +142,7 @@ public class OrientDBEntityQueries extends QueriesBase {
 	 * @return list of associations
 	 */
 
-	public List<Map<String, Object>> findAssociation(Connection connection, AssociationKey associationKey,
+	public List<Map<String, Object>> findAssociation(ODatabaseDocumentTx db, AssociationKey associationKey,
 			AssociationContext associationContext) {
 		List<Map<String, Object>> association = new LinkedList<>();
 		log.debugf( "findAssociation: associationKey: %s; associationContext: %s", associationKey, associationContext );
@@ -191,37 +167,32 @@ public class OrientDBEntityQueries extends QueriesBase {
 		}
 
 		log.debugf( "findAssociation: query: %s", query );
-		try {
-			Statement stmt = connection.createStatement();
-			ResultSet rs = stmt.executeQuery( query.toString() );
-			while ( rs.next() ) {
-				Map<String, Object> dbValues = new LinkedHashMap<>();
-				ResultSetMetaData metadata = rs.getMetaData();
-				for ( String systemField : OrientDBConstant.SYSTEM_FIELDS ) {
-					dbValues.put( systemField, rs.getObject( systemField ) );
+		List<ODocument> documents = QueryUtil.executeNativeQuery( db, query );
+		for ( ODocument doc : documents ) {
+			Map<String, Object> dbValues = doc.toMap();
+			for ( String fieldName : dbValues.keySet() ) {
+				if ( dbValues.get( fieldName ) instanceof ODocument ) {
+					dbValues.put( fieldName, ODocumentUtil.extractNamesTree( fieldName, (ODocument) dbValues.get( fieldName ) ) );
 				}
-				for ( int i = 0; i < rs.getMetaData().getColumnCount(); i++ ) {
-					int dbFieldNo = i + 1;
-					String dbColumnName = metadata.getColumnName( dbFieldNo );
-					if ( isLinkedProperty( dbColumnName ) ) {
-						continue;
-					}
-					Object dbValue = rs.getObject( dbColumnName );
-					dbValues.put( dbColumnName, dbValue );
-					if ( dbValue instanceof ODocument ) {
-						dbValues.remove( dbColumnName );
-						dbValues.putAll( ODocumentUtil.extractNamesTree( dbColumnName, (ODocument) dbValue ) );
-					}
-				}
-				log.debugf( " entiry values from db: %s", dbValues );
-				association.add( dbValues );
 			}
-			log.debugf( "findAssociation: rows :  %d", association.size() );
-		}
-		catch (SQLException sqle) {
-			throw log.cannotExecuteQuery( query.toString(), sqle );
 
+			association.add( dbValues );
 		}
+		log.debugf( "findAssociation: rows :  %d", association.size() );
+
+		/*
+		 * try { Statement stmt = connection.createStatement(); ResultSet rs = stmt.executeQuery( query.toString() );
+		 * while ( rs.next() ) { Map<String, Object> dbValues = new LinkedHashMap<>(); ResultSetMetaData metadata =
+		 * rs.getMetaData(); for ( String systemField : OrientDBConstant.SYSTEM_FIELDS ) { dbValues.put( systemField,
+		 * rs.getObject( systemField ) ); } for ( int i = 0; i < rs.getMetaData().getColumnCount(); i++ ) { int
+		 * dbFieldNo = i + 1; String dbColumnName = metadata.getColumnName( dbFieldNo ); if ( isLinkedProperty(
+		 * dbColumnName ) ) { continue; } Object dbValue = rs.getObject( dbColumnName ); dbValues.put( dbColumnName,
+		 * dbValue ); if ( dbValue instanceof ODocument ) { dbValues.remove( dbColumnName ); dbValues.putAll(
+		 * ODocumentUtil.extractNamesTree( dbColumnName, (ODocument) dbValue ) ); } } log.debugf(
+		 * " entiry values from db: %s", dbValues ); association.add( dbValues ); } log.debugf(
+		 * "findAssociation: rows :  %d", association.size() ); } catch (SQLException sqle) { throw
+		 * log.cannotExecuteQuery( query.toString(), sqle ); }
+		 */
 		return association;
 	}
 }
