+++ orientdb/src/main/java/org/hibernate/ogm/datastore/orientdb/query/impl/FloatParamValueSetter.java
@@ -6,7 +6,6 @@
  */
 package org.hibernate.ogm.datastore.orientdb.query.impl;
 
-import java.sql.PreparedStatement;
 import java.sql.SQLException;
 
 /**
@@ -17,8 +16,8 @@ import java.sql.SQLException;
 public class FloatParamValueSetter implements ParamValueSetter<Float> {
 
 	@Override
-	public void setValue(PreparedStatement preparedStatement, int index, Float value) throws SQLException {
-		preparedStatement.setFloat( index, value );
+	public void setValue(int index, Float value) throws SQLException {
+		// preparedStatement.setFloat( index, value );
 	}
 
 }
