+++ orientdb/src/main/java/org/hibernate/ogm/datastore/orientdb/query/impl/ShortParamValueSetter.java
@@ -6,7 +6,6 @@
  */
 package org.hibernate.ogm.datastore.orientdb.query.impl;
 
-import java.sql.PreparedStatement;
 import java.sql.SQLException;
 
 /**
@@ -17,8 +16,8 @@ import java.sql.SQLException;
 public class ShortParamValueSetter implements ParamValueSetter<Short> {
 
 	@Override
-	public void setValue(PreparedStatement preparedStatement, int index, Short value) throws SQLException {
-		preparedStatement.setShort( index, value );
+	public void setValue(int index, Short value) throws SQLException {
+		// preparedStatement.setShort( index, value );
 	}
 
 }
