+++ orientdb/src/test/java/org/hibernate/ogm/datastore/orientdb/test/OrientDBTest.java
@@ -4,11 +4,13 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.ogm.datastore.orientdb.test.jpa;
+package org.hibernate.ogm.datastore.orientdb.test;
+
+import org.hibernate.ogm.backendtck.type.BuiltInTypeTest;
 
 /**
- * @author Sergey Chernolyas &lt;sergey.chernolyas@gmail.com&gt;
+ * @author Davide D'Alto
  */
-public enum Status {
-	USUAL, VIP;
+public class OrientDBTest extends BuiltInTypeTest {
+
 }
