+++ orientdb/src/test/java/org/hibernate/ogm/datastore/orientdb/test/jpa/OrientDbEmbeddedTest.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.ogm.datastore.orientdb.test;
+package org.hibernate.ogm.datastore.orientdb.test.jpa;
 
 import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
@@ -16,20 +16,17 @@ import java.util.LinkedList;
 import java.util.List;
 
 import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
-import javax.persistence.FlushModeType;
-import javax.persistence.Persistence;
 
 import org.apache.log4j.Logger;
-import org.hibernate.ogm.datastore.orientdb.test.jpa.Car;
-import org.hibernate.ogm.datastore.orientdb.test.jpa.CarOwner;
-import org.hibernate.ogm.datastore.orientdb.test.jpa.EngineInfo;
-import org.hibernate.ogm.datastore.orientdb.test.jpa.Producer;
-import org.hibernate.ogm.datastore.orientdb.utils.MemoryDBUtil;
+import org.hibernate.ogm.datastore.orientdb.test.jpa.entity.Car;
+import org.hibernate.ogm.datastore.orientdb.test.jpa.entity.CarOwner;
+import org.hibernate.ogm.datastore.orientdb.test.jpa.entity.Customer;
+import org.hibernate.ogm.datastore.orientdb.test.jpa.entity.EngineInfo;
+import org.hibernate.ogm.datastore.orientdb.test.jpa.entity.Pizza;
+import org.hibernate.ogm.datastore.orientdb.test.jpa.entity.Producer;
+import org.hibernate.ogm.utils.jpa.OgmJpaTestCase;
 import org.junit.After;
-import org.junit.AfterClass;
 import org.junit.Before;
-import org.junit.BeforeClass;
 import org.junit.FixMethodOrder;
 import org.junit.Test;
 import org.junit.runners.MethodSorters;
@@ -38,40 +35,21 @@ import org.junit.runners.MethodSorters;
  * @author Sergey Chernolyas &lt;sergey.chernolyas@gmail.com&gt;
  */
 @FixMethodOrder(MethodSorters.NAME_ASCENDING)
-public class OrientDbEmbeddedTest {
-
-	public static final String MEMORY_TEST = "memory:test";
-	private static final Logger log = Logger.getLogger( OrientDBSimpleTest.class.getName() );
-	private static EntityManager em;
-	private static EntityManagerFactory emf;
-
-	@BeforeClass
-	public static void setUpClass() {
-		MemoryDBUtil.createDbFactory( MEMORY_TEST );
-		emf = Persistence.createEntityManagerFactory( "hibernateOgmJpaUnit" );
-		em = emf.createEntityManager();
-		em.setFlushMode( FlushModeType.COMMIT );
-	}
-
-	@AfterClass
-	public static void tearDownClass() {
-		if ( em != null ) {
-			em.close();
-			emf.close();
-
-		}
-		MemoryDBUtil.dropInMemoryDb();
-	}
+public class OrientDbEmbeddedTest extends OgmJpaTestCase {
+	private static final Logger log = Logger.getLogger( OrientDbEmbeddedTest.class.getName() );
+	private EntityManager em;
 
 	@Before
 	public void setUp() {
-		if ( em.getTransaction().isActive() ) {
-			em.getTransaction().rollback();
-		}
+		em = getFactory().createEntityManager();
+
 	}
 
 	@After
 	public void tearDown() {
+		if ( em.getTransaction().isActive() ) {
+			em.getTransaction().rollback();
+		}
 		em.clear();
 	}
 
@@ -225,5 +203,9 @@ public class OrientDbEmbeddedTest {
 			throw e;
 		}
 	}
-
+	@Override
+	protected Class<?>[] getAnnotatedClasses() {
+		return new Class<?>[] { Car.class,CarOwner.class,Customer.class,
+			EngineInfo.class, Pizza.class, Producer.class };
+	}
 }
