+++ orientdb/src/test/java/org/hibernate/ogm/datastore/orientdb/test/jpa/entity/PassportPK.java
@@ -4,25 +4,18 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.ogm.datastore.orientdb.test.jpa;
+package org.hibernate.ogm.datastore.orientdb.test.jpa.entity;
 
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.IdClass;
+import java.io.Serializable;
 
 /**
  * @author Sergey Chernolyas &lt;sergey.chernolyas@gmail.com&gt;
  */
-@Entity
-@IdClass(PassportPK.class)
-public class Passport {
 
-	@Id
+public class PassportPK implements Serializable {
+
 	private int seria;
-	@Id
 	private long number;
-
-	private String fio;
 
 	public int getSeria() {
 		return seria;
@@ -40,19 +33,11 @@ public class Passport {
 		this.number = number;
 	}
 
-	public String getFio() {
-		return fio;
-	}
-
-	public void setFio(String fio) {
-		this.fio = fio;
-	}
-
 	@Override
 	public int hashCode() {
-		int hash = 3;
-		hash = 19 * hash + this.seria;
-		hash = 19 * hash + (int) ( this.number ^ ( this.number >>> 32 ) );
+		int hash = 7;
+		hash = 23 * hash + this.seria;
+		hash = 23 * hash + (int) ( this.number ^ ( this.number >>> 32 ) );
 		return hash;
 	}
 
@@ -67,7 +52,7 @@ public class Passport {
 		if ( getClass() != obj.getClass() ) {
 			return false;
 		}
-		final Passport other = (Passport) obj;
+		final PassportPK other = (PassportPK) obj;
 		if ( this.seria != other.seria ) {
 			return false;
 		}
@@ -75,11 +60,6 @@ public class Passport {
 			return false;
 		}
 		return true;
-	}
-
-	@Override
-	public String toString() {
-		return "Passport{" + "seria=" + seria + ", number=" + number + ", fio=" + fio + '}';
 	}
 
 }
