+++ orientdb/src/test/java/org/hibernate/ogm/datastore/orientdb/test/jpa/entity/Producer.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.ogm.datastore.orientdb.test.jpa;
+package org.hibernate.ogm.datastore.orientdb.test.jpa.entity;
 
 import javax.persistence.Embeddable;
