+++ orientdb/src/test/java/org/hibernate/ogm/datastore/orientdb/test/jpa/entity/Product.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.ogm.datastore.orientdb.test.jpa;
+package org.hibernate.ogm.datastore.orientdb.test.jpa.entity;
 
 import java.util.List;
 import java.util.Objects;
@@ -12,12 +12,11 @@ import java.util.Objects;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.ManyToMany;
-import javax.persistence.OneToMany;
 import javax.persistence.Version;
 
+import org.hibernate.annotations.GenericGenerator;
 import org.hibernate.search.annotations.Indexed;
 
 import com.orientechnologies.orient.core.id.ORecordId;
@@ -26,19 +25,16 @@ import com.orientechnologies.orient.core.id.ORecordId;
  * @author Sergey Chernolyas &lt;sergey.chernolyas@gmail.com&gt;
  */
 @Entity
-@Indexed(index = "Pizza")
-public class Pizza {
+@Indexed(index = "Product")
+public class Product {
 
 	@Id
-	@Column(name = "bKey")
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long bKey;
+	@GeneratedValue(generator = "uuid")
+	@GenericGenerator(name = "uuid", strategy = "uuid2")
+	private String id;
 	private String name;
-	@OneToMany(mappedBy = "buying")
-	private List<OrderItem> orderItems;
-
-	@ManyToMany
-	private List<Product> products;
+	@ManyToMany(mappedBy = "products")
+	private List<Pizza> pizzas;
 
 	@Version
 	@Column(name = "@version")
@@ -46,12 +42,12 @@ public class Pizza {
 	@Column(name = "@rid")
 	private ORecordId rid;
 
-	public Long getbKey() {
-		return bKey;
+	public String getId() {
+		return id;
 	}
 
-	public void setbKey(Long bKey) {
-		this.bKey = bKey;
+	public void setId(String id) {
+		this.id = id;
 	}
 
 	public String getName() {
@@ -62,12 +58,12 @@ public class Pizza {
 		this.name = name;
 	}
 
-	public List<OrderItem> getOrderItems() {
-		return orderItems;
+	public List<Pizza> getPizzas() {
+		return pizzas;
 	}
 
-	public void setOrderItems(List<OrderItem> orderItems) {
-		this.orderItems = orderItems;
+	public void setPizzas(List<Pizza> pizzas) {
+		this.pizzas = pizzas;
 	}
 
 	public int getVersion() {
@@ -86,18 +82,10 @@ public class Pizza {
 		this.rid = rid;
 	}
 
-	public List<Product> getProducts() {
-		return products;
-	}
-
-	public void setProducts(List<Product> products) {
-		this.products = products;
-	}
-
 	@Override
 	public int hashCode() {
 		int hash = 7;
-		hash = 13 * hash + Objects.hashCode( this.bKey );
+		hash = 29 * hash + Objects.hashCode( this.id );
 		return hash;
 	}
 
@@ -112,11 +100,10 @@ public class Pizza {
 		if ( getClass() != obj.getClass() ) {
 			return false;
 		}
-		final Pizza other = (Pizza) obj;
-		if ( !Objects.equals( this.bKey, other.bKey ) ) {
+		final Product other = (Product) obj;
+		if ( !Objects.equals( this.id, other.id ) ) {
 			return false;
 		}
 		return true;
 	}
-
 }
