+++ orientdb/src/test/java/org/hibernate/ogm/datastore/orientdb/test/jpa/entity/Status.java
@@ -4,13 +4,11 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.ogm.datastore.orientdb;
-
-import org.hibernate.ogm.backendtck.type.BuiltInTypeTest;
+package org.hibernate.ogm.datastore.orientdb.test.jpa.entity;
 
 /**
- * @author Davide D'Alto
+ * @author Sergey Chernolyas &lt;sergey.chernolyas@gmail.com&gt;
  */
-public class OrientDBTest extends BuiltInTypeTest {
-
+public enum Status {
+	USUAL, VIP;
 }
