+++ orientdb/src/test/java/org/hibernate/ogm/datastore/orientdb/test/jpa/entity/Writer.java
@@ -4,66 +4,44 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.ogm.datastore.orientdb.test.jpa;
-
-import java.util.List;
-import java.util.Objects;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.ManyToMany;
-import javax.persistence.Version;
-
-import org.hibernate.annotations.GenericGenerator;
-import org.hibernate.search.annotations.Indexed;
+package org.hibernate.ogm.datastore.orientdb.test.jpa.entity;
 
 import com.orientechnologies.orient.core.id.ORecordId;
+import java.util.Date;
+import java.util.Objects;
+import javax.persistence.Column;
+import javax.persistence.Convert;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.Temporal;
+import javax.persistence.TemporalType;
+import javax.persistence.Version;
 
 /**
  * @author Sergey Chernolyas &lt;sergey.chernolyas@gmail.com&gt;
  */
 @Entity
-@Indexed(index = "Product")
-public class Product {
+public class Writer {
 
 	@Id
-	@GeneratedValue(generator = "uuid")
-	@GenericGenerator(name = "uuid", strategy = "uuid2")
-	private String id;
-	private String name;
-	@ManyToMany(mappedBy = "products")
-	private List<Pizza> pizzas;
-
+	private Long bKey;
 	@Version
 	@Column(name = "@version")
 	private int version;
 	@Column(name = "@rid")
 	private ORecordId rid;
+	@Temporal(TemporalType.DATE)
+	private Date birthDate;
+	@Convert(converter = StringToUpperCaseConverter.class)
+	private String name;
+	private long count;
 
-	public String getId() {
-		return id;
+	public Long getbKey() {
+		return bKey;
 	}
 
-	public void setId(String id) {
-		this.id = id;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	public List<Pizza> getPizzas() {
-		return pizzas;
-	}
-
-	public void setPizzas(List<Pizza> pizzas) {
-		this.pizzas = pizzas;
+	public void setbKey(Long bKey) {
+		this.bKey = bKey;
 	}
 
 	public int getVersion() {
@@ -82,10 +60,34 @@ public class Product {
 		this.rid = rid;
 	}
 
+	public Date getBirthDate() {
+		return birthDate;
+	}
+
+	public void setBirthDate(Date birthDate) {
+		this.birthDate = birthDate;
+	}
+
+	public String getName() {
+		return name;
+	}
+
+	public void setName(String name) {
+		this.name = name;
+	}
+
+	public long getCount() {
+		return count;
+	}
+
+	public void setCount(long count) {
+		this.count = count;
+	}
+
 	@Override
 	public int hashCode() {
-		int hash = 7;
-		hash = 29 * hash + Objects.hashCode( this.id );
+		int hash = 5;
+		hash = 29 * hash + Objects.hashCode( this.bKey );
 		return hash;
 	}
 
@@ -100,10 +102,11 @@ public class Product {
 		if ( getClass() != obj.getClass() ) {
 			return false;
 		}
-		final Product other = (Product) obj;
-		if ( !Objects.equals( this.id, other.id ) ) {
+		final Writer other = (Writer) obj;
+		if ( !Objects.equals( this.bKey, other.bKey ) ) {
 			return false;
 		}
 		return true;
 	}
+
 }
