+++ orientdb/src/test/java/org/hibernate/ogm/datastore/orientdb/test/jpa/entity/package-info.java
@@ -1,0 +1,7 @@
+/*
+ * Hibernate OGM, Domain model persistence for NoSQL datastores
+ *
+ * License: GNU Lesser General Public License (LGPL), version 2.1 or later
+ * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
+ */
+package org.hibernate.ogm.datastore.orientdb.test.jpa.entity;
